+++ src/camera.h
@@ -16,7 +16,7 @@ public:
     const int height, width;
     Camera(glm::vec3 pos, glm::vec3 target, glm::vec3 up,
            float fov_y, int width, int height);
-    glm::vec3 getRayDir(int i, int j) const;
+    glm::vec3 getRayDir(float i, float j) const;
     glm::vec3 getPos() { return pos; }
 };
